requirements.txt
openpyxl
pandas
plotly
psycopg2-binary
python-dotenv
SQLAlchemy
streamlit